bot Aroon/requirements.txt
# DEPENDENCIAS PRINCIPALES
pandas>=2.1.3
ta==0.10.2
requests>=2.32.3
websockets==11.0.3
numpy>=1.26.0,<2.0.0

# TELEGRAM BOT
python-telegram-bot==20.7

# ANALISIS TECNICO ADICIONAL
scipy>=1.11.4
scikit-learn>=1.3.2

# UTILIDADES
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1

# LOGGING Y MONITOREO
loguru==0.7.2

# FECHAS Y TIEMPO
pytz==2023.3

# CRIPTOGRAFIA
cryptography==41.0.7

# QUOTEX API (requiere Python >= 3.12)
pyquotex @ git+https://github.com/cleitonleonel/pyquotex.git

# ANÁLISIS DE DATOS
matplotlib==3.8.2
seaborn==0.13.0

# ASYNC SUPPORT
aiofiles==23.2.1
aiohttp==3.9.1

# PDF GENERATION
reportlab==4.0.7